Extract the shared Quasar cmd-buf definitions into noc_cmd_buf_common.h
Pure code motion out of noc_nonblocking_api_v2.h: register-layout
constants and wrappers, MISC/VC values, counter bookkeeping,
command-buffer initialization, and the flush/barrier predicates. The
upcoming NOC V3 API shares the same RoCC transport generation and will
include this header instead of duplicating ~700 lines of V2.

Object code verified byte-identical for the Quasar kernel and firmware
probes and the full kernel sweep.
